Moved style transformations to a cloned node
This commit solves #14

I have been analysing the behaviour of stretchy with CSS transitions,
and the problem starts on the "computed styles". When the width or
height values are being animated, the computed value changes do not
occur immediately, and this is why the calculated value is not correct.

Firstly I tried to queue the calculation until the end of the
transition by using setInterval, but the UI felt buggy and flickery, so
I decided to move these transformation to a new node, clone of the
original element, and only apply the result at the end of the process.
This cloned element needs to be added to the DOM, as the browser
computes the changed values, but it can be removed when everything is
calculated.

Aaaand that's all. I know that creating a node, appending it to the DOM,
and removing it may look too heavy, but I have tested in on Chrome and
Firefox and the results are pretty good :)